+++ frontend/src/bff/operations/remove-product.js
@@ -1,21 +1,18 @@
-// import { deleteComment, deleteProduct, getComments } from '../api';
-// import { sessions } from '../sessions';
-// import { ROLE } from '../constants';
-
 import { deleteProduct } from '../api';
+import { ROLE } from '../constants';
+import { sessions } from '../sessions';
 
-export const removeProduct = async (productId) => {
-    // export const removeProduct = async (hash, id) => {
-    // const accessRoles = [ROLE.ADMIN];
+export const removeProduct = async (hash, productId) => {
+    const accessRoles = [ROLE.ADMIN];
 
-    // const access = await sessions.access(hash, accessRoles);
+    const access = await sessions.access(hash, accessRoles);
 
-    // if (!access) {
-    //     return {
-    //         error: 'Доступ запрещен',
-    //         res: null,
-    //     };
-    // }
+    if (!access) {
+        return {
+            error: 'Доступ запрещен',
+            res: null,
+        };
+    }
 
     await deleteProduct(productId);
